[Refact] callType domain으로 이동
