--- data/src/main/java/com/yapp/buddycon/data/datasource/remote/GetGiftConListDataSourceImpl.kt
@@ -1,6 +1,0 @@
-package com.yapp.buddycon.data.datasource.remote
-
-import com.yapp.buddycon.domain.model.CouponInfo
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
-
