+++ data/src/main/java/com/yapp/buddycon/data/repository/CouponRepositoryImpl.kt
@@ -4,8 +4,9 @@ import androidx.paging.*
 import com.yapp.buddycon.data.datasource.local.giftcon.CouponRemoteMediator
 import com.yapp.buddycon.data.db.BuddyConDataBase
 import com.yapp.buddycon.data.network.api.CouponService
-import com.yapp.buddycon.domain.model.CouponInfo
 import com.yapp.buddycon.domain.model.CouponItem
+import com.yapp.buddycon.domain.model.CouponType
+import com.yapp.buddycon.domain.model.SortMode
 import com.yapp.buddycon.domain.repository.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
