+++ domain/src/main/java/com/yapp/buddycon/domain/model/CallType.kt
@@ -1,0 +1,16 @@
+package com.yapp.buddycon.domain.model
+
+enum class TabMode {
+    Usable, Used, Custom
+}
+
+enum class SortMode(val value: String) {
+    NoShared("noshared"),
+    ExpireDate("expireDate"),
+    Name("name"),
+    CreatedAt("createdAt")
+}
+
+enum class CouponType {
+    GiftCon, Custom, Made
+}
