+++ domain/src/main/java/com/yapp/buddycon/domain/repository/CouponRepository.kt
@@ -2,18 +2,9 @@ package com.yapp.buddycon.domain.repository
 
 import androidx.paging.PagingData
 import com.yapp.buddycon.domain.model.CouponItem
+import com.yapp.buddycon.domain.model.CouponType
+import com.yapp.buddycon.domain.model.SortMode
 import kotlinx.coroutines.flow.Flow
-
-enum class SortMode(val value: String) {
-    NoShared("noshared"),
-    ExpireDate("expireDate"),
-    Name("name"),
-    CreatedAt("createdAt")
-}
-
-enum class CouponType {
-    GiftCon, Custom, Made
-}
 
 interface CouponRepository {
     fun getCouponList(
