--- domain/src/main/java/com/yapp/buddycon/domain/repository/GetGiftConListRepository.kt
@@ -1,8 +1,0 @@
-package com.yapp.buddycon.domain.repository
-
-import com.yapp.buddycon.domain.model.CouponInfo
-import kotlinx.coroutines.flow.Flow
-
-interface GetGiftConListRepository {
-    fun getGiftConList(page : Int, size : Int, sort : List<String>) : Flow<List<CouponInfo>>
-}
