+++ domain/src/main/java/com/yapp/buddycon/domain/usecase/giftcon/GetCouponInfoUseCase.kt
@@ -1,11 +1,10 @@
 package com.yapp.buddycon.domain.usecase.giftcon
 
 import androidx.paging.PagingData
-import com.yapp.buddycon.domain.model.CouponInfo
 import com.yapp.buddycon.domain.model.CouponItem
+import com.yapp.buddycon.domain.model.CouponType
+import com.yapp.buddycon.domain.model.SortMode
 import com.yapp.buddycon.domain.repository.CouponRepository
-import com.yapp.buddycon.domain.repository.CouponType
-import com.yapp.buddycon.domain.repository.SortMode
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
