+++ presentation/src/main/java/com/yapp/buddycon/presentation/ui/giftcon/GiftConFragment.kt
@@ -8,12 +8,12 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.GridLayoutManager
-import com.yapp.buddycon.domain.repository.SortMode
+import com.yapp.buddycon.domain.model.SortMode
+import com.yapp.buddycon.domain.model.TabMode
 import com.yapp.buddycon.presentation.R
 import com.yapp.buddycon.presentation.base.BaseFragment
 import com.yapp.buddycon.presentation.databinding.FragmentGiftconBinding
 import com.yapp.buddycon.presentation.ui.main.BuddyConViewModel
-import com.yapp.buddycon.presentation.ui.main.TabMode
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.launchIn
